rename indexd_test_utils2 -> pytest_indexd
