+++ .secrets.baseline
@@ -153,16 +153,6 @@
         "is_secret": false
       }
     ],
-    "indexd_test_utils2/__init__.py": [
-      {
-        "type": "Secret Keyword",
-        "filename": "indexd_test_utils2/__init__.py",
-        "hashed_secret": "d033e22ae348aeb5660fc2140aec35850c4da997",
-        "is_verified": false,
-        "line_number": 89,
-        "is_secret": false
-      }
-    ],
     "psql-users.sh": [
       {
         "type": "Secret Keyword",
@@ -170,6 +160,16 @@
         "hashed_secret": "a94a8fe5ccb19ba61c4c0873d391e987982fbbd3",
         "is_verified": false,
         "line_number": 1,
+        "is_secret": false
+      }
+    ],
+    "pytest_indexd/__init__.py": [
+      {
+        "type": "Secret Keyword",
+        "filename": "pytest_indexd/__init__.py",
+        "hashed_secret": "d033e22ae348aeb5660fc2140aec35850c4da997",
+        "is_verified": false,
+        "line_number": 89,
         "is_secret": false
       }
     ],
@@ -208,5 +208,5 @@
       }
     ]
   },
-  "generated_at": "2022-12-24T17:43:11Z"
+  "generated_at": "2023-01-03T16:07:48Z"
 }
