+++ tests/integration/conftest.py
@@ -4,9 +4,7 @@ import os
 
 import pytest
 
-if os.getenv("USE_TEST_UTILS2", "false").lower() == "true":
-    from indexd_test_utils2.plugin import *
-else:
+if os.getenv("USE_TEST_UTILS2", "false").lower() == "false":
     from indexd_test_utils import (
         alias_driver,
         auth_driver,
